[Service] Link to existing third party when creating a service
When we create a new service, the website may already be in the database
as a third party.  In this case, we want to establish a link to the
existing third party, to ensure it can share metadata.  This case will
be relatively rare, but possible, so it is explicitly added here.  The
reverse direction (linking new third party hosts to existing services)
will be added next.