[API] Add Bookmarklet API and JS
The new API app contains a tiny API needed for a bookmarklet that
easily creates new identities to sign them up for a newsletter.

api/views.py@BookmarkletApiView contains the code the bookmarklet talks
to, and the bookmarklet itself (in a comment).
identity/static/js/bookmarklet/bookmarklet.js contains the javascript
code that is loaded and executed by the bookmarklet.

This API is currently unauthenticated and not rate limited. This should
probably be changed for the final production version (or the bookmarklet
could just be removed again, although it may be handy to have).