Update cronjob to only run once
The cronjob becomes unstable after a while. This commit changes the cron to only run one cycle of retrieve => view => click => analyze instead of many. This may slow down the process overall as it leads to more idle time, but if it works around the bug in OpenWPM that leads to the crashes, this will very much be worth it.